Change license from GPLv2 to Mozilla Public License 2.0
hlulsl.nsi was the only file to have an explicit GPLv2 license, and
I was the only one to have contributed code to that. Several other
scripts had contributions by others (thank you!), but I have
determined they've been refactored out over the years and do not
need to be reverted / reimplemented.